+++ app/[locale]/projects/[projectId]/fundings/__tests__/page.spec.tsx
@@ -1,44 +1,27 @@
 import React from 'react';
-import { act, render, screen, fireEvent, waitFor } from '@testing-library/react';
+import { render, screen, fireEvent, waitFor } from '@testing-library/react';
 import { useRouter, useParams } from 'next/navigation';
 import '@testing-library/jest-dom';
-import { MockedProvider } from '@apollo/client/testing';
 import { axe, toHaveNoViolations } from 'jest-axe';
 
-import { ProjectFundingsDocument } from '@/generated/graphql';
-
 import ProjectsProjectFunding from '../page';
+import { useProjectFundingsQuery } from '@/generated/graphql';
+import { useToast } from '@/context/ToastContext';
 
+import mockFunders from '../__mocks__/mockFunders.json';
+
+
+
+jest.mock("@/generated/graphql", () => ({
+  useProjectFundingsQuery: jest.fn(),
+}));
 
 expect.extend(toHaveNoViolations);
 
 const mockPush = jest.fn();
-
-const mocks = [
-  {
-    request: {
-      query: ProjectFundingsDocument,
-      variables: {
-        projectId: 123,
-      },
-    },
-
-    result: {
-      data: {
-        projectFundings: [
-          {
-            id: 1,
-            affiliation: {
-              displayName: "Test Funder 1",
-              uri: "https://funder-1",
-            },
-          }
-        ],
-      },
-    }
-  },
-];
-
+const mockToast = {
+  add: jest.fn(),
+};
 
 describe('ProjectsProjectFunding', () => {
   beforeEach(() => {
@@ -48,39 +31,49 @@ describe('ProjectsProjectFunding', () => {
 
     const mockParams = useParams as jest.Mock;
     mockParams.mockReturnValue({ projectId: '123' });
+
+    // Mock Toast
+    (useToast as jest.Mock).mockReturnValue(mockToast);
+
+    // Mock the hook for data state
+    (useProjectFundingsQuery as jest.Mock).mockReturnValue({
+      data: mockFunders,
+      loading: false,
+      error: null,
+    });
   })
 
   afterEach(() => {
     jest.clearAllMocks();
   });
 
-  it('should render the page header with title and description', () => {
+  it('should render the page header with title and description', async () => {
     render(
-      <MockedProvider mocks={mocks} addTypename={false}>
         <ProjectsProjectFunding />
-      </MockedProvider>
     );
 
     expect(screen.getByText('Project Funding')).toBeInTheDocument();
     expect(screen.getByText('Manage funding sources for your project')).toBeInTheDocument();
+    // breadcrumbs
+    expect(screen.getByText('breadcrumbs.home')).toBeInTheDocument();
+    expect(screen.getByText('breadcrumbs.projects')).toBeInTheDocument();
+    expect(screen.getByText('breadcrumbs.projectOverview')).toBeInTheDocument();
+    expect(screen.getByText('breadcrumbs.projectFunding')).toBeInTheDocument();
   });
 
-  it('should render breadcrumbs correctly', () => {
+  it('should render breadcrumbs correctly', async () => {
     render(
-      <MockedProvider mocks={mocks} addTypename={false}>
-        <ProjectsProjectFunding />
-      </MockedProvider>
+      <ProjectsProjectFunding />
     );
 
-    expect(screen.getByText('Home')).toBeInTheDocument();
-    expect(screen.getByText('Projects')).toBeInTheDocument();
+    expect(screen.getByText('breadcrumbs.home')).toBeInTheDocument();
+    expect(screen.getByText('breadcrumbs.projectOverview')).toBeInTheDocument();
+    expect(screen.getByText('breadcrumbs.projectFunding')).toBeInTheDocument();
   });
 
   it('should render the "Add funding" button and handles click', async () => {
     render(
-      <MockedProvider mocks={mocks} addTypename={false}>
-        <ProjectsProjectFunding />
-      </MockedProvider>
+      <ProjectsProjectFunding />
     );
 
     const addButton = screen.getByRole('button', { name: 'Add funding' });
@@ -94,30 +87,40 @@ describe('ProjectsProjectFunding', () => {
   });
 
   it('should render the fundings list and handles "Edit" button click', async () => {
-    act(() => {
-      render(
-        <MockedProvider mocks={mocks} addTypename={false}>
-          <ProjectsProjectFunding />
-        </MockedProvider>
-      );
-    });
+    render(
+      <ProjectsProjectFunding />
+    );
 
     await waitFor(() => {
-      const editButton = screen.getByLabelText('Edit Test Funder 1 details');
+      const editButton = screen.getByLabelText('Edit Irish Research Council (research.ie) details');
       expect(editButton).toBeInTheDocument();
       fireEvent.click(editButton);
     });
 
     await waitFor(() => {
-      expect(mockPush).toHaveBeenCalledWith('/en-US/projects/123/fundings/projFund_6902/edit');
+      expect(mockPush).toHaveBeenCalledWith('/en-US/projects/123/fundings/IRL-000000X1/edit');
+    });
+  });
+
+  it('should display toast error when a user clicks a funder without a funderProjectNumber', async () => {
+    render(
+      <ProjectsProjectFunding />
+    );
+
+    await waitFor(() => {
+      const editButton = screen.getByLabelText('Edit National Science Foundation (nsf.gov) details');
+      expect(editButton).toBeInTheDocument();
+      fireEvent.click(editButton);
+    });
+
+    await waitFor(() => {
+      expect(mockToast.add).toHaveBeenCalledWith('messages.errors.funderNumberNotFound', { type: 'error' });
     });
   });
 
   it('should pass accessibility tests', async () => {
     const { container } = render(
-      <MockedProvider mocks={mocks} addTypename={false}>
         <ProjectsProjectFunding />
-      </MockedProvider>
     );
     const results = await axe(container);
     expect(results).toHaveNoViolations();