+++ app/[locale]/projects/[projectId]/page.tsx
@@ -59,7 +59,7 @@ const ProjectOverviewPage: React.FC = () => {
     plans: [],
     projectMembers: [],
     researchOutputs: []
-  });
+  });  
 
   // Localization keys
   const ProjectOverview = useTranslations('ProjectOverview');
@@ -142,10 +142,10 @@ const ProjectOverviewPage: React.FC = () => {
         showBackButton={false}
         breadcrumbs={
           <Breadcrumbs aria-label={ProjectOverview('navigation')}>
-            <Breadcrumb><Link href="/">{ProjectOverview('home')}</Link></Breadcrumb>
-            <Breadcrumb><Link href="/projects">{ProjectOverview('projects')}</Link></Breadcrumb>
-            <Breadcrumb>{ProjectOverview('pageTitle')}</Breadcrumb>
-          </Breadcrumbs>
+            <Breadcrumb><Link href={routePath('app.home')}>{ProjectOverview('home')}</Link></Breadcrumb>
+            <Breadcrumb><Link href={routePath('projects.index')}>{ProjectOverview('projects')}</Link></Breadcrumb>
+            <Breadcrumb>{Global('breadcrumbs.projectOverview')}</Breadcrumb>  
+          </Breadcrumbs>   
         }
         actions={null}
         className="page-project-list"